refactor(web): extract shared article helpers
