+++ apps/web/src/components/articles/shared-article-view.tsx
@@ -45,6 +45,20 @@ import {
   useToggleSharedArticleEmpathy,
   useUpdateSharedArticleComment,
 } from "@/lib/api/articles";
+import {
+  type CommentSortOption,
+  type ViewerProfile,
+  extractViewerProfile,
+  formatCommentDate,
+  formatPublishedDate,
+  insertCommentTree,
+  normalizeDisplayName,
+  removeCommentTree,
+  sortCommentThread,
+  toCommentSortOption,
+  updateCommentTree,
+  withThreadDefaults,
+} from "@/components/articles/shared-article-helpers";
 import { apiClient } from "@/lib/api-client";
 import { getSupabase } from "@/lib/auth/auth-client";
 import { locales } from "@/lib/i18n/config";
@@ -106,241 +120,6 @@ function LocaleSwitcher() {
 
 function isUuidLike(value: string): boolean {
   return UUID_LIKE_RE.test(value);
-}
-
-interface ViewerProfile {
-  id: string | null;
-  name: string | null;
-  image: string | null;
-}
-
-type CommentSortOption = "latest" | "recommended";
-
-const CONTENT_TYPE_STYLES: Record<string, { labelKey: ContentTypeLabelKey; className: string }> = {
-  tech_blog: {
-    labelKey: "typeTechBlog",
-    className:
-      "border-blue-200 bg-blue-50 text-blue-700 dark:border-blue-400/40 dark:bg-blue-500/15 dark:text-blue-200",
-  },
-  academic_paper: {
-    labelKey: "typePaper",
-    className:
-      "border-purple-200 bg-purple-50 text-purple-700 dark:border-purple-400/40 dark:bg-purple-500/15 dark:text-purple-200",
-  },
-  general_news: {
-    labelKey: "typeNews",
-    className:
-      "border-gray-200 bg-gray-50 text-gray-700 dark:border-gray-400/40 dark:bg-gray-500/15 dark:text-gray-200",
-  },
-  github_repo: {
-    labelKey: "typeGitHub",
-    className:
-      "border-slate-200 bg-slate-50 text-slate-700 dark:border-slate-400/40 dark:bg-slate-500/15 dark:text-slate-200",
-  },
-  official_docs: {
-    labelKey: "typeDocs",
-    className:
-      "border-teal-200 bg-teal-50 text-teal-700 dark:border-teal-400/40 dark:bg-teal-500/15 dark:text-teal-200",
-  },
-  video_podcast: {
-    labelKey: "typeVideo",
-    className:
-      "border-pink-200 bg-pink-50 text-pink-700 dark:border-pink-400/40 dark:bg-pink-500/15 dark:text-pink-200",
-  },
-  patch_note: {
-    labelKey: "typePatchNote",
-    className:
-      "border-amber-200 bg-amber-50 text-amber-700 dark:border-amber-400/40 dark:bg-amber-500/15 dark:text-amber-200",
-  },
-};
-
-type ContentTypeLabelKey =
-  | "typeTechBlog"
-  | "typePaper"
-  | "typeNews"
-  | "typeGitHub"
-  | "typeDocs"
-  | "typeVideo"
-  | "typePatchNote";
-
-function readMetadataValue(metadata: unknown, key: string): string | null {
-  if (typeof metadata !== "object" || metadata === null) {
-    return null;
-  }
-
-  const value = (metadata as Record<string, unknown>)[key];
-  if (typeof value !== "string") {
-    return null;
-  }
-
-  const trimmed = value.trim();
-  return trimmed.length > 0 ? trimmed : null;
-}
-
-function extractViewerProfile(user: unknown): ViewerProfile | null {
-  if (typeof user !== "object" || user === null) {
-    return null;
-  }
-
-  const userRecord = user as Record<string, unknown>;
-  const id =
-    typeof userRecord.id === "string" && userRecord.id.trim().length > 0 ? userRecord.id : null;
-  const email =
-    typeof userRecord.email === "string" && userRecord.email.trim().length > 0
-      ? userRecord.email
-      : null;
-  const metadata = userRecord.user_metadata;
-
-  const name =
-    readMetadataValue(metadata, "full_name") ?? readMetadataValue(metadata, "name") ?? email;
-  const image = readMetadataValue(metadata, "avatar_url") ?? readMetadataValue(metadata, "picture");
-
-  return {
-    id,
-    name,
-    image,
-  };
-}
-
-function formatPublishedDate(value: string, locale: string): string | null {
-  const date = new Date(value);
-  if (Number.isNaN(date.getTime())) {
-    return null;
-  }
-
-  return new Intl.DateTimeFormat(locale, {
-    year: "numeric",
-    month: "2-digit",
-    day: "2-digit",
-  }).format(date);
-}
-
-function withThreadDefaults(comments: SharedArticleComment[]): SharedArticleComment[] {
-  return comments.map((comment) => ({
-    ...comment,
-    author_image: comment.author_image ?? null,
-    parent_comment_id: comment.parent_comment_id ?? null,
-    empathy_count: comment.empathy_count ?? 0,
-    viewer_has_empathy: comment.viewer_has_empathy ?? false,
-    replies: withThreadDefaults(comment.replies ?? []),
-  }));
-}
-
-function updateCommentTree(
-  comments: SharedArticleComment[],
-  targetCommentId: string,
-  updater: (comment: SharedArticleComment) => SharedArticleComment
-): SharedArticleComment[] {
-  return comments.map((comment) => {
-    if (comment.id === targetCommentId) {
-      return updater(comment);
-    }
-
-    if (comment.replies.length === 0) {
-      return comment;
-    }
-
-    return {
-      ...comment,
-      replies: updateCommentTree(comment.replies, targetCommentId, updater),
-    };
-  });
-}
-
-function insertCommentTree(
-  comments: SharedArticleComment[],
-  newComment: SharedArticleComment
-): SharedArticleComment[] {
-  if (!newComment.parent_comment_id) {
-    return [newComment, ...comments];
-  }
-
-  return updateCommentTree(comments, newComment.parent_comment_id, (comment) => ({
-    ...comment,
-    replies: [...comment.replies, newComment],
-  }));
-}
-
-function removeCommentTree(
-  comments: SharedArticleComment[],
-  targetCommentId: string
-): SharedArticleComment[] {
-  return comments.flatMap((comment) => {
-    if (comment.id === targetCommentId) {
-      return [];
-    }
-
-    if (comment.replies.length === 0) {
-      return [comment];
-    }
-
-    return [
-      {
-        ...comment,
-        replies: removeCommentTree(comment.replies, targetCommentId),
-      },
-    ];
-  });
-}
-
-function formatCommentDate(
-  value: string,
-  locale: string
-): { label: string; dateTime: string } | null {
-  const date = new Date(value);
-  if (Number.isNaN(date.getTime())) {
-    return null;
-  }
-
-  const now = new Date();
-  const yearOptions = now.getFullYear() === date.getFullYear() ? {} : { year: "numeric" as const };
-
-  return {
-    label: new Intl.DateTimeFormat(locale, {
-      month: "short",
-      day: "numeric",
-      ...yearOptions,
-    }).format(date),
-    dateTime: date.toISOString(),
-  };
-}
-
-function normalizeDisplayName(value: string | null | undefined): string {
-  return (value ?? "").trim().toLowerCase();
-}
-
-function toCommentSortOption(value: string): CommentSortOption {
-  return value === "recommended" ? "recommended" : "latest";
-}
-
-function getCommentTimestamp(comment: SharedArticleComment): number {
-  const timestamp = Date.parse(comment.created_at);
-  return Number.isNaN(timestamp) ? 0 : timestamp;
-}
-
-function getCommentEmpathyCount(comment: SharedArticleComment): number {
-  return typeof comment.empathy_count === "number" ? comment.empathy_count : 0;
-}
-
-function sortCommentThread(
-  comments: SharedArticleComment[],
-  sortOption: CommentSortOption
-): SharedArticleComment[] {
-  return [...comments]
-    .sort((left, right) => {
-      if (sortOption === "recommended") {
-        const empathyDiff = getCommentEmpathyCount(right) - getCommentEmpathyCount(left);
-        if (empathyDiff !== 0) {
-          return empathyDiff;
-        }
-      }
-
-      return getCommentTimestamp(right) - getCommentTimestamp(left);
-    })
-    .map((comment) => ({
-      ...comment,
-      replies: sortCommentThread(comment.replies, sortOption),
-    }));
 }
 
 export function SharedArticleView({
